requirements.txt
flask
flask-cors
joblib
numpy
pandas
scikit-learn
sentence-transformers
transformers
torch
PyPDF2
python-docx
gunicorn
